Separate directory for storage implementations
