+++ packages/powersync_core/lib/attachments/attachments.dart
@@ -7,6 +7,6 @@ library;
 
 export '../src/attachments/attachment.dart';
 export '../src/attachments/attachment_queue_service.dart';
-export '../src/attachments/local_storage.dart';
+export '../src/attachments/storage/local_storage.dart';
 export '../src/attachments/remote_storage.dart';
 export '../src/attachments/sync_error_handler.dart';
