+++ packages/powersync_core/lib/src/attachments/sync/syncing_service.dart
@@ -7,7 +7,7 @@ import 'package:async/async.dart';
 import '../attachment.dart';
 import '../implementations/attachment_context.dart';
 import '../implementations/attachment_service.dart';
-import '../local_storage.dart';
+import '../storage/local_storage.dart';
 import '../remote_storage.dart';
 import '../sync_error_handler.dart';
