+++ packages/powersync_core/test/attachments/local_storage_test.dart
@@ -7,7 +7,7 @@ import 'dart:typed_data';
 
 import 'package:test/test.dart';
 import 'package:path/path.dart' as p;
-import 'package:powersync_core/src/attachments/io_local_storage.dart';
+import 'package:powersync_core/src/attachments/storage/io_local_storage.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
 
 void main() {
